Add note about method calling order to readme
This documents a problem I've run into repeatedly, and I finally figured out the cause, specifically in the case of `nullable`. Because the `nullable` property is used in `attach` to add a `nullable` validation rule, the validation will fail if `nullable()` has not been called before `attach`. Looking at the code in `attach`, it's likely this applies to various methods such as `options` and `displayUsing`.